+++ pytorch_lightning/callbacks/lr_monitor.py
@@ -19,6 +19,7 @@ Learning Rate Monitor
 Monitor and logs learning rate for lr schedulers during training.
 
 """
+import itertools
 from collections import defaultdict
 from typing import Any, DefaultDict, Dict, List, Optional, Set, Tuple, Type
 
@@ -123,7 +124,7 @@ class LearningRateMonitor(Callback):
                 )
 
         # Find names for schedulers
-        names: List[str] = []
+        names: List[List[str]] = []
         (
             sched_hparam_keys,
             optimizers_with_scheduler,
@@ -140,8 +141,9 @@ class LearningRateMonitor(Callback):
         names.extend(optimizer_hparam_keys)
 
         # Initialize for storing values
-        self.lrs = {name: [] for name in names}
-        self.last_momentum_values = {name + "-momentum": None for name in names}
+        names_flatten = list(itertools.chain.from_iterable(names))
+        self.lrs = {name: [] for name in names_flatten}
+        self.last_momentum_values = {name + "-momentum": None for name in names_flatten}
 
     def on_train_batch_start(self, trainer: "pl.Trainer", *args: Any, **kwargs: Any) -> None:
         if not trainer.logger_connector.should_update_logs:
@@ -172,7 +174,7 @@ class LearningRateMonitor(Callback):
         ) = self._find_names_from_schedulers(trainer.lr_schedulers, add_lr_sch_names=False)
         self._remap_keys(scheduler_hparam_keys)
 
-        for name, scheduler in zip(self.lr_sch_names, trainer.lr_schedulers):
+        for name, scheduler in zip(scheduler_hparam_keys, trainer.lr_schedulers):
             if interval in [scheduler["interval"], "any"]:
                 opt = scheduler["scheduler"].optimizer
                 current_stat = self._get_lr_momentum_stat(opt, name)
@@ -186,23 +188,22 @@ class LearningRateMonitor(Callback):
         )
         self._remap_keys(optimizer_hparam_keys)
 
-        for opt, name in zip(optimizers_without_scheduler, optimizer_hparam_keys):
-            current_stat = self._get_lr_momentum_stat(opt, name)
+        for opt, names in zip(optimizers_without_scheduler, optimizer_hparam_keys):
+            current_stat = self._get_lr_momentum_stat(opt, names)
             latest_stat.update(current_stat)
 
         return latest_stat
 
-    def _get_lr_momentum_stat(self, optimizer: Optimizer, name: str) -> Dict[str, float]:
+    def _get_lr_momentum_stat(self, optimizer: Optimizer, names: List[str]) -> Dict[str, float]:
         lr_momentum_stat = {}
         param_groups = optimizer.param_groups
         use_betas = "betas" in optimizer.defaults
 
-        for i, pg in enumerate(param_groups):
-            name_and_suffix = self._add_suffix(name, param_groups, i)
-            lr = self._extract_lr(pg, name_and_suffix)
+        for pg, name in zip(param_groups, names):
+            lr = self._extract_lr(pg, name)
             lr_momentum_stat.update(lr)
             momentum = self._extract_momentum(
-                param_group=pg, name=name_and_suffix.replace(name, f"{name}-momentum"), use_betas=use_betas
+                param_group=pg, name=name.replace(name, f"{name}-momentum"), use_betas=use_betas
             )
             lr_momentum_stat.update(momentum)
 
@@ -213,14 +214,15 @@ class LearningRateMonitor(Callback):
         self.lrs[name].append(lr)
         return {name: lr}
 
-    def _remap_keys(self, names: List[str], token: str = "/pg1") -> None:
+    def _remap_keys(self, names: List[List[str]], token: str = "/pg1") -> None:
         """This function is used the remap the keys if param groups for a given optimizer increased."""
-        for new_name in names:
-            old_name = new_name.replace(token, "")
-            if token in new_name and old_name in self.lrs:
-                self.lrs[new_name] = self.lrs.pop(old_name)
-            elif new_name not in self.lrs:
-                self.lrs[new_name] = []
+        for group_new_names in names:
+            for new_name in group_new_names:
+                old_name = new_name.replace(token, "")
+                if token in new_name and old_name in self.lrs:
+                    self.lrs[new_name] = self.lrs.pop(old_name)
+                elif new_name not in self.lrs:
+                    self.lrs[new_name] = []
 
     def _extract_momentum(self, param_group: Dict[str, List], name: str, use_betas: bool) -> Dict[str, float]:
         if not self.log_momentum:
@@ -258,7 +260,7 @@ class LearningRateMonitor(Callback):
 
     def _find_names_from_schedulers(
         self, lr_schedulers: List, add_lr_sch_names: bool = True
-    ) -> Tuple[List[str], List[Optimizer], DefaultDict[Type[Optimizer], int]]:
+    ) -> Tuple[List[List[str]], List[Optimizer], DefaultDict[Type[Optimizer], int]]:
         # Create unique names in the case we have multiple of the same learning
         # rate scheduler + multiple parameter groups
         names = []
@@ -271,10 +273,11 @@ class LearningRateMonitor(Callback):
             else:
                 name = "lr-" + sch.optimizer.__class__.__name__
 
-            updated_name = self._check_duplicates_and_update_name(
+            updated_names = self._check_duplicates_and_update_name(
                 sch.optimizer, name, seen_optimizers, seen_optimizer_types, scheduler, add_lr_sch_names
             )
-            names.extend(updated_name)
+            names.append(updated_names)
+
         return names, seen_optimizers, seen_optimizer_types
 
     def _find_names_from_optimizers(
@@ -283,7 +286,7 @@ class LearningRateMonitor(Callback):
         seen_optimizers: List[Optimizer],
         seen_optimizer_types: DefaultDict[Type[Optimizer], int],
         add_lr_sch_names: bool = True,
-    ) -> Tuple[List[str], List[Optimizer]]:
+    ) -> Tuple[List[List[str]], List[Optimizer]]:
         names = []
         optimizers_without_scheduler = []
 
@@ -294,11 +297,12 @@ class LearningRateMonitor(Callback):
                 continue
 
             name = "lr-" + optimizer.__class__.__name__
-            updated_name = self._check_duplicates_and_update_name(
+            updated_names = self._check_duplicates_and_update_name(
                 optimizer, name, seen_optimizers, seen_optimizer_types, None, add_lr_sch_names
             )
-            names.extend(updated_name)
+            names.append(updated_names)
             optimizers_without_scheduler.append(optimizer)
+
         return names, optimizers_without_scheduler
 
     def _check_duplicates_and_update_name(